project/training/requirements.txt
torch==2.0.1
torchvision==0.15.2
opencv-python==4.8.0.74
Pillow==10.0.0
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
scikit-learn==1.3.0